halo profiles in their own directory
