+++ pyccl/halos/profiles/__init__.py
@@ -1,0 +1,20 @@
+from .profile_base import HaloProfile
+from .gaussian import HaloProfileGaussian
+from .powerlaw import HaloProfilePowerLaw
+from .nfw import HaloProfileNFW
+from .einasto import HaloProfileEinasto
+from .hernquist import HaloProfileHernquist
+from .pressure_gnfw import HaloProfilePressureGNFW
+from .hod import HaloProfileHOD
+
+
+__all__ = (
+    "HaloProfile",
+    "HaloProfileGaussian",
+    "HaloProfilePowerLaw",
+    "HaloProfileNFW",
+    "HaloProfileEinasto",
+    "HaloProfileHernquist",
+    "HaloProfilePressureGNFW",
+    "HaloProfileHOD",
+)
